trunc_string_at is moved to util/analytics_platform_util.py
