+++ analytics_platform/kronos/pgm/src/pgm_util.py
@@ -20,16 +20,3 @@ def generate_matrix_from_pandas_df(df, node_list):
     y = np.vstack((x, np.ones((1, len(node_list)))))
     result = np.vstack((y, np.zeros((1, len(node_list)))))
     return result
-
-
-def trunc_string_at(s, d, n1, n2):
-    """
-    Returns s truncated at the n'th occurrence of the delimiter, d
-    """
-    if n2 > 0:
-        result = d.join(s.split(d, n2)[n1:n2])
-    else:
-        result = d.join(s.split(d, n2)[n1:])
-        if not result.endswith("/"):
-            result += "/"
-    return result
