+++ analytics_platform/kronos/src/kronos_online_scoring.py
@@ -247,50 +247,3 @@ def score_eco_user_package_dict(user_request, user_eco_kronos_dict, eco_to_krono
 
         response_json_list.append(prediction_result_dict)
     return response_json_list
-
-
-if __name__ == "__main__":
-    from analytics_platform.kronos.pgm.src.offline_training import load_eco_to_kronos_dependency_dict_s3
-
-    training_url = "s3://perf-gsk-data/python/machine-learning"
-
-    bucket_name = trunc_string_at(training_url, "/", 2, 3)
-    additional_path = trunc_string_at(training_url, "/", 3, -1)
-
-    user_eco_kronos_dict = load_user_eco_to_kronos_model_dict_s3(bucket_name=bucket_name,
-                                                                 additional_path=additional_path)
-
-    print user_eco_kronos_dict
-
-    eco_to_kronos_dependency_dict = load_eco_to_kronos_dependency_dict_s3(bucket_name=bucket_name,
-                                                                          additional_path=additional_path)
-
-    print "loading from s3 done\n"
-
-    input_json = [
-        {
-            "ecosystem": "pypi",
-            "comp_package_count_threshold": 5,
-            "alt_package_count_threshold": 3,
-            "outlier_probability_threshold": 0.61,
-            "unknown_packages_ratio_threshold": 0.2,
-            "package_list": [
-                "click",
-                "httpie",
-                "parsimonious",
-                "prompt-toolkit",
-                "pygments",
-                "wheel",
-                "setuptools",
-                "pydispatcher",
-                "haha",
-                "huhu"
-            ]
-        }
-    ]
-
-    response = score_eco_user_package_dict(user_request=input_json, user_eco_kronos_dict=user_eco_kronos_dict,
-                                           eco_to_kronos_dependency_dict=eco_to_kronos_dependency_dict)
-    import json
-
-    print json.dumps(response)
